Move WelcomeMessage classe into a DirectMessages module.
