+++ lib/twitter.rb
@@ -29,7 +29,7 @@ require 'twitter/target_user'
 require 'twitter/trend'
 require 'twitter/tweet'
 require 'twitter/user'
-require 'twitter/welcome_message'
-require 'twitter/welcome_message_rule'
-require 'twitter/welcome_message_rule_wrapper'
-require 'twitter/welcome_message_wrapper'
+require 'twitter/direct_messages/welcome_message'
+require 'twitter/direct_messages/welcome_message_rule'
+require 'twitter/direct_messages/welcome_message_rule_wrapper'
+require 'twitter/direct_messages/welcome_message_wrapper'
