+++ lib/twitter/direct_messages/welcome_message_rule.rb
@@ -1,0 +1,12 @@
+require 'twitter/creatable'
+require 'twitter/identity'
+
+module Twitter
+  module DirectMessages
+    class WelcomeMessageRule < Twitter::Identity
+      include Twitter::Creatable
+      # @return [Integer]
+      attr_reader :welcome_message_id
+    end
+  end
+end
