+++ lib/twitter/rest/direct_messages/welcome_messages.rb
@@ -22,7 +22,7 @@ module Twitter
             }
           }
           json_options[:welcome_message][:name] = name if name
-          event = perform_request_with_object(:json_post, '/1.1/direct_messages/welcome_messages/new.json', json_options.merge!(options), Twitter::WelcomeMessageWrapper)
+          event = perform_request_with_object(:json_post, '/1.1/direct_messages/welcome_messages/new.json', json_options.merge!(options), Twitter::DirectMessages::WelcomeMessageWrapper)
           event.welcome_message
         end
 
@@ -39,20 +39,20 @@ module Twitter
               text: text
             }
           }
-          event = perform_request_with_object(:json_put, '/1.1/direct_messages/welcome_messages/update.json', json_options.merge!(options), Twitter::WelcomeMessageWrapper, params)
+          event = perform_request_with_object(:json_put, '/1.1/direct_messages/welcome_messages/update.json', json_options.merge!(options), Twitter::DirectMessages::WelcomeMessageWrapper, params)
           event.welcome_message
         end
 
         def welcome_message(id, options = {})
           options = options.dup
           options[:id] = id
-          event = perform_get_with_object('/1.1/direct_messages/welcome_messages/show.json', options, Twitter::WelcomeMessageWrapper)
+          event = perform_get_with_object('/1.1/direct_messages/welcome_messages/show.json', options, Twitter::DirectMessages::WelcomeMessageWrapper)
           event.welcome_message
         end
 
         def welcome_message_list(options = {})
           limit = options.fetch(:count, 20)
-          events = perform_get_with_cursor('/1.1/direct_messages/welcome_messages/list.json', options.merge!(no_default_cursor: true, count: 50, limit: limit), :welcome_messages, Twitter::WelcomeMessageWrapper)
+          events = perform_get_with_cursor('/1.1/direct_messages/welcome_messages/list.json', options.merge!(no_default_cursor: true, count: 50, limit: limit), :welcome_messages, Twitter::DirectMessages::WelcomeMessageWrapper)
           events.collect(&:welcome_message)
         end
 
@@ -64,7 +64,7 @@ module Twitter
                 welcome_message_id: welcome_message_id
             }
           }
-          event = perform_request_with_object(:json_post, '/1.1/direct_messages/welcome_messages/rules/new.json', json_options.merge!(options), Twitter::WelcomeMessageRuleWrapper)
+          event = perform_request_with_object(:json_post, '/1.1/direct_messages/welcome_messages/rules/new.json', json_options.merge!(options), Twitter::DirectMessages::WelcomeMessageRuleWrapper)
           event.welcome_message_rule
         end
 
@@ -75,13 +75,13 @@ module Twitter
         def welcome_message_rule(id, options = {})
           options = options.dup
           options[:id] = id
-          event = perform_get_with_object('/1.1/direct_messages/welcome_messages/rules/show.json', options, Twitter::WelcomeMessageRuleWrapper)
+          event = perform_get_with_object('/1.1/direct_messages/welcome_messages/rules/show.json', options, Twitter::DirectMessages::WelcomeMessageRuleWrapper)
           event.welcome_message_rule
         end
 
         def welcome_message_rule_list(options = {})
           limit = options.fetch(:count, 20)
-          events = perform_get_with_cursor('/1.1/direct_messages/welcome_messages/rules/list.json', options.merge!(no_default_cursor: true, count: 50, limit: limit), :welcome_message_rules, Twitter::WelcomeMessageRuleWrapper)
+          events = perform_get_with_cursor('/1.1/direct_messages/welcome_messages/rules/list.json', options.merge!(no_default_cursor: true, count: 50, limit: limit), :welcome_message_rules, Twitter::DirectMessages::WelcomeMessageRuleWrapper)
           events.collect(&:welcome_message_rule)
         end
       end
