--- lib/twitter/welcome_message_rule.rb
@@ -1,10 +1,0 @@
-require 'twitter/creatable'
-require 'twitter/identity'
-
-module Twitter
-  class WelcomeMessageRule < Twitter::Identity
-    include Twitter::Creatable
-    # @return [Integer]
-    attr_reader :welcome_message_id
-  end
-end
